refactor: get rid of barrel files
