--- src/api/consent/index.ts
@@ -1,1 +1,0 @@
-export * from "./consent";
