--- src/api/custom-map/index.ts
@@ -1,1 +1,0 @@
-export * from "./custom-map";
