--- src/api/ecommerce/index.ts
@@ -1,1 +1,0 @@
-export * from "./ecommerce";
