--- src/api/event/index.ts
@@ -1,1 +1,0 @@
-export * from "./event";
