--- src/api/exception/index.ts
@@ -1,1 +1,0 @@
-export * from "./exception";
