--- src/api/index.ts
@@ -1,13 +1,0 @@
-export * from "./config";
-export * from "./consent";
-export * from "./custom-map";
-export * from "./ecommerce";
-export * from "./event";
-export * from "./exception";
-export * from "./linker";
-export * from "./opt";
-export * from "./pageview";
-export * from "./query";
-export * from "./screenview";
-export * from "./set";
-export * from "./time";
