--- src/api/set/index.ts
@@ -1,1 +1,0 @@
-export * from "./set";
