+++ src/api/time.ts
@@ -1,0 +1,6 @@
+import { event } from "@/api/event";
+import type { GtagEventParams } from "@/types/gtag";
+
+export function time(params: GtagEventParams) {
+  event("timing_complete", params);
+}
