--- src/api/time/index.ts
@@ -1,1 +1,0 @@
-export * from "./time";
