--- src/api/time/time.ts
@@ -1,6 +1,0 @@
-import type { GtagEventParams } from "../../types/gtag";
-import { event } from "../event";
-
-export function time(params: GtagEventParams) {
-  event("timing_complete", params);
-}
