+++ src/core/add-configuration.ts
