--- src/core/add-configuration/index.ts
@@ -1,1 +1,0 @@
-export * from "./add-configuration";
