--- src/core/add-gtag/index.ts
@@ -1,1 +1,0 @@
-export * from "./add-gtag";
