--- src/core/add-router-tracking/index.ts
@@ -1,1 +1,0 @@
-export * from "./add-router-tracking";
