+++ src/core/create-gtag.ts
@@ -1,10 +1,19 @@
-import * as api from "@/api/index";
+import { config } from "@/api/config";
+import { consent, consentDeniedAll, consentGrantedAll } from "@/api/consent";
+import { customMap } from "@/api/custom-map";
+import { ecommerce } from "@/api/ecommerce";
+import { event } from "@/api/event";
+import { exception } from "@/api/exception";
+import { linker } from "@/api/linker";
+import { optIn, optOut } from "@/api/opt";
+import { pageview } from "@/api/pageview";
+import { query } from "@/api/query";
+import { screenview } from "@/api/screenview";
+import { set } from "@/api/set";
+import { time } from "@/api/time";
 import { addGtag } from "@/core/add-gtag";
-import {
-  type PluginSettings,
-  getSettings,
-  updateSettings,
-} from "@/core/settings";
+import { getSettings, updateSettings } from "@/core/settings";
+import type { PluginSettings } from "@/types/settings";
 import type { App } from "vue";
 
 function initGtag() {
@@ -25,6 +34,25 @@ export function configure(settings: PluginSettings): void {
   updateSettings(settings);
   initGtag();
 }
+
+const api = {
+  config,
+  consent,
+  consentDeniedAll,
+  consentGrantedAll,
+  customMap,
+  ecommerce,
+  event,
+  exception,
+  linker,
+  optIn,
+  optOut,
+  pageview,
+  screenview,
+  set,
+  time,
+  query,
+};
 
 type GtagAPI = typeof api;
