--- src/core/create-gtag/index.ts
@@ -1,1 +1,0 @@
-export * from "./create-gtag";
