--- src/core/index.ts
@@ -1,2 +1,0 @@
-export * from "./add-gtag";
-export * from "./create-gtag";
