--- src/core/settings/index.ts
@@ -1,2 +1,0 @@
-export * from "./settings";
-export * from "./types";
