+++ src/core/track-route.ts
@@ -1,10 +1,7 @@
 import { pageview } from "@/api/pageview";
 import { screenview } from "@/api/screenview";
-import {
-  type PageTrackerParams,
-  type Route,
-  getSettings,
-} from "@/core/settings";
+import { getSettings } from "@/core/settings";
+import type { PageTrackerParams, Route } from "@/types/settings";
 
 function isRouteExcluded(route: Route): boolean {
   const { pageTracker } = getSettings();
