+++ src/tests/core/create-gtag.test.ts
@@ -1,10 +1,41 @@
-import * as api from "@/api/index";
+import { config } from "@/api/config";
+import { consent, consentDeniedAll, consentGrantedAll } from "@/api/consent";
+import { customMap } from "@/api/custom-map";
+import { ecommerce } from "@/api/ecommerce";
+import { event } from "@/api/event";
+import { exception } from "@/api/exception";
+import { linker } from "@/api/linker";
+import { optIn, optOut } from "@/api/opt";
+import { pageview } from "@/api/pageview";
+import { query } from "@/api/query";
+import { screenview } from "@/api/screenview";
+import { set } from "@/api/set";
+import { time } from "@/api/time";
 import { addGtag } from "@/core/add-gtag";
 import { createGtag } from "@/core/create-gtag";
 import { resetSettings } from "@/core/settings";
 import { createApp } from "vue";
 
 vi.mock("@/core/add-gtag");
+
+const api = {
+  config,
+  consent,
+  consentDeniedAll,
+  consentGrantedAll,
+  customMap,
+  ecommerce,
+  event,
+  exception,
+  linker,
+  optIn,
+  optOut,
+  pageview,
+  screenview,
+  set,
+  time,
+  query,
+};
 
 describe("createGtag", () => {
   beforeEach(resetSettings);
