+++ src/types/settings.ts
